% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netFun.R
\name{netFun}
\alias{netFun}
\title{netFun}
\usage{
netFun(x, ...)
}
\arguments{
\item{x}{Numeric matrix with samples in columns.}
}
\description{
This is the network reconstruction function that will be used to build aggregate networks.
}
\examples{
netFun()
}
\keyword{netFun}
