% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lioness.R
\name{lioness}
\alias{lioness}
\title{LIONESS}
\usage{
lioness(x, f = netFun, assay = 'counts')
}
\arguments{
\item{x}{Numeric matrix with samples in columns or SummerizedExperiment object}

\item{f}{Network reconstruction function. Defaults to Pearson correlation.}
\item{assay}{Specified the assay used in lionessR for SummerizedExperiment object. Default: counts}
}
\description{
This function uses the LIONESS equation to estimate single-sample networks.
}
\value{
Dataframe of single-sample association network.
}
\examples{
exp <- matrix(sample(1000,1000)/1000, 100, 10)
colnames(exp) <- paste("sample", c(1:ncol(exp)), sep="_")
row.names(exp) <- paste("gene", c(1:nrow(exp)), sep="_")
lionessResults <- lioness(exp, netFun)
}
\keyword{lioness}
