% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netFun.R
\name{netFun}
\alias{netFun}
\title{netFun}
\usage{
netFun(x)
}
\arguments{
\item{x}{Numeric matrix with samples in columns.}
}
\description{
This is the network reconstruction function that will be used to build aggregate networks.
}
\value{
Numeric matrix of association network. 
}
\examples{
exp <- matrix(sample(1000, 1000)/1000, 100, 10)
netFun(exp)
}
\keyword{netFun}
